+++ packages/mq_remote_client/lib/mq_remote_client.dart
@@ -1,6 +1,6 @@
 /// MQ remote client package
 library mq_remote_client;
 
-export 'src/http_exception.dart';
+export 'src/exception/mq_remote_exception.dart';
 export 'src/mq_remote_client.dart';
-export 'src/network_client.dart';
+export 'src/network_client/network_client.dart';
