+++ packages/mq_remote_client/lib/src/network_client/network_client.dart
@@ -1,5 +1,5 @@
 import 'package:connectivity_plus/connectivity_plus.dart';
-import 'package:flutter/foundation.dart';
+import 'package:meta/meta.dart';
 
 /// A client for checking the device's internet connection status.
 @immutable
